Extract applied cards component
