+++ src/lib/checkout/checkout.js
@@ -5,6 +5,7 @@ import {
   CHECKOUT_STEP_SHIPPING_METHOD
 } from "../constants";
 import { CheckoutForm } from "./checkout_form";
+import { CheckoutAppliedCards } from "./checkout_applied_cards";
 
 const SUPPORTED_CHECKOUT_STEPS = [
   CHECKOUT_STEP_CONTACT_INFORMATION,
@@ -36,13 +37,20 @@ export class Checkout {
 
     // define an event handler for page changes
     const handlePageChange = () => {
-      document.querySelectorAll(SELECTOR_CHECKOUT_REDUCTION_FORM_WRAPPER).forEach(formWrapperElement => {
+      document.querySelectorAll(SELECTOR_CHECKOUT_REDUCTION_FORM_WRAPPER).forEach((formWrapperElement, index) => {
         // skip if the form wrapper is already initialised
         if(formWrapperElement.dataset.cardigan === 'true') {
           return;
         }
 
+        // initialise the checkout form component on this element
         new CheckoutForm(formWrapperElement, api, config, templates);
+
+        // if this is the first element on the page, also initialise the applied cards component
+        // this component should only exist once on the page
+        if(index === 0) {
+          new CheckoutAppliedCards(formWrapperElement, api, config);
+        }
       });
     };
